apollo: overlay: Don't let any carrier default WFC mode to `IMS Preferred`
Qualcomm has added another WFC preference setting `IMS Preferred`
which is not supported in AOSP. If we let carriers default to this
setting, the AOSP settings app shows that WFC is disabled and
`Unexpected WFC mode value: 10` is logged. Even though the default
mode is still working fine, the user is unable to change to this
mode again once a different mode was selected. To fix this, remove
all default WFC mode settings with value "10" and let these carriers
default to "2" (`WiFi Preferred`, the default in AOSP) instead.

Leave the `carrier_wfc_supports_ims_preferred_bool` configs in place,
such that, in case the following patches are present in the system,
the user is still able to select the `IMS Preferred` mode in settings
manually.

https://git.codelinaro.org/clo/la/platform/frameworks/base/-/commit/097dacd55bf88a0baff805e5d21bc52aafca1b1e
https://git.codelinaro.org/clo/la/platform/packages/services/Telephony/-/commit/a451b1fd2305622a79ed42bf23627cb47b1e3c08
https://git.codelinaro.org/clo/la/platform/packages/apps/Settings/-/commit/a79d737742902e730090222589ecf4c19fa1b024
https://git.codelinaro.org/clo/la/platform/packages/apps/Settings/-/commit/d8ff55888f59f6340dc7a14f9ab44bb2a9d17fce#
https://git.codelinaro.org/clo/la/platform/packages/apps/Settings/-/commit/a2beaccd8fb78baf49da3ed0948b6cca45b5bd17

Change-Id: Ie6b9f30da45450c43a465de1170ef11d1d4ba1ee
Signed-off-by: Gustavo Mendes <gusttavo.me@outlook.com>